pyproject.toml
[project]
name = "python-task-scheduler"
version = "0.1.0"
description = "Python 定时任务调度服务 - 独立的定时任务管理应用"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
